Divide classes into packages, all classes are under duke package
