--- src/main/java/ListCommand.java
@@ -1,12 +1,0 @@
-public class ListCommand extends Command {
-
-    @Override
-    public void execute(TaskList taskList, Ui ui, Storage storage) {
-        ui.showList(taskList);
-    }
-
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
