+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.exception.InvalidTaskIndexException;
+import duke.task.Task;
+import duke.task.TaskList;
+
 public class DeleteCommand extends Command {
     private int taskIndex;
