+++ src/main/java/duke/command/TodayCommand.java
@@ -1,0 +1,21 @@
+package duke.command;
+
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
+import java.time.LocalDate;
+
+public class TodayCommand extends Command {
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        LocalDate today = LocalDate.now();
+        ui.showTasksForToday(today, taskList);
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
