+++ src/main/java/duke/exception/EmptyCommandException.java
@@ -1,6 +1,8 @@
+package duke.exception;
+
 public class EmptyCommandException extends DukeException {
 
     public EmptyCommandException() {
-        super(" ☹ OOPS!!! Please enter a valid command.");
+        super(" ☹ OOPS!!! Please enter a valid duke.command.");
     }
 }
