+++ src/main/java/duke/exception/InvalidFormatException.java
@@ -1,6 +1,8 @@
+package duke.exception;
+
 public class InvalidFormatException extends DukeException {
 
     public InvalidFormatException(String formatDescription) {
-        super(" ☹ OOPS!!! The format of the command is invalid. \n\t " + formatDescription);
+        super(" ☹ OOPS!!! The format of the duke.command is invalid. \n\t " + formatDescription);
     }
 }
