+++ src/main/java/duke/exception/InvalidTaskIndexException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class InvalidTaskIndexException extends DukeException {
 
     public InvalidTaskIndexException(int taskIndex) {
