+++ src/main/java/duke/exception/UnknownCommandException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class UnknownCommandException extends DukeException {
     public UnknownCommandException() {
         super(" ☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
